000052 - Put French keypad on admin page
Moved the controller code which displays the French character map into a trait called CharMapTrait.php.

Refactored QuestionController.php and question/index.phtml to use the trait.

Using the trait, added the character map to the "add" and "edit" pages of the admin page.

Modified the javascript in charmap.phtml to move the focus to the answer field after appending the character.